Created interfaces for functions, added speed testing, reworked overall structure
