+++ data_input.h
@@ -1,0 +1,6 @@
+#ifndef DATAINPUT_H
+#define DATAINPUT_H
+
+void DataInput();
+
+#endif
