+++ test_speed.h
@@ -1,0 +1,6 @@
+#ifndef TESTSPEED_H
+#define TESTSPEED_H
+
+void TestSpeed();
+
+#endif
